`overlapToMatrix()` generic
setup `overlapToMatrix()` as a generic.
The giotto object and lower methods do have some repairs that need to be done on the matrix in order for it to fit into the object. This might need to be revised in the future since it makes it so that compatible matrix generation is not fully encapsulated in the lower methods.