Fix subset testthat
added toplevel_params to subsetGiottoLocs with a value of 5 so that it no longer tries to look for .subset_giotto() which is an internal function.

This has always been an issue since subsetGiottoLocs was broken up into several functions, but it only threw an error after the functions for grabbing previous function characteristics were moved to GiottoUtils which cannot access GiottoClass internals.