fix!: `.evaluate_expr_matrix` --- Breaking change: arg 'dbSparseMatrix' changed to 'dbMatrix' to support both dbDenseMatrix and dbSparseMatrix.
This is a temporary patch to this internal function to be able to pass in dbMatrix without supplying a constructor. As is, this function is inflexible with passing the full set of args to the constructor for each alt Matrix object.  Future rewrite by @jiajic will coerce matrix logic through s3/s4 method dispatch and provide a more flexible interface.